requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# LLM libraries
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
openai>=1.0.0
anthropic>=0.18.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pydantic>=2.0.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
httpx>=0.25.0

# Optional - for notebook
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

